Create sum_and_difference_of_two_numbers.c
OBJECTIVE- 

The fundamental data types in c are int, float and char. Today, we're discussing int and float data types.
The printf() function prints the given statement to the console. The syntax is printf("format string",argument_list);. In the function, if we are using an integer, character, string or float as argument, then in the format string we have to write %d (integer), %c (character), %s (string), %f (float) respectively.
The scanf() function reads the input data from the console. The syntax is scanf("format string",argument_list);. For ex: The scanf("%d",&number) statement reads integer number from the console and stores the given value in variable .
To input two integers separated by a space on a single line, the command is scanf("%d %d", &n, &m), where  and  are the two integers.

---------------------------------------------------------------------------------------------------------

Sample Input

10 4
4.0 2.0

Sample Output

14 6
6.0 2.0