Create functions_in_c.c
Objective-

In this challenge, you will learn simple usage of functions in C. Functions are a bunch of statements grouped together. A function is provided with zero or more arguments, and it executes the statements on it. Based on the return type, it either returns nothing (void) or something.
--------------------------------------------------------------------------------------------------------------------------------------------

Sample Input

3
4
6
5

Sample Output

6